extract ngram hash to a seperate file

+++ src/config.h
@@ -25,6 +25,8 @@
 #ifndef  _CONNLM_CONFIG_H_
 #define  _CONNLM_CONFIG_H_
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -51,7 +53,7 @@ typedef unsigned long count_t;
 #define COUNT_FMT "%lu"
 #define COUNT_MAX ((count_t)-1)
 
-typedef unsigned long long hash_t;
+typedef uint64_t hash_t;
 #define HASH_FMT "%llu"
 
 #define ALIGN_SIZE 128
